# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/suppliers": {
            "get": {
                "description": "Show a supplier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Suppliers"
                ],
                "summary": "Show supplier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Supplier identification",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ShowSupplierResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a supplier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Suppliers"
                ],
                "summary": "Update supplier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Supplier Identification",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Supplier data to Update",
                        "name": "supplier",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateSupplierRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateSupplierResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new supplier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Suppliers"
                ],
                "summary": "Create Supplier",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateSupplierRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.CreateSupplierResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a new supplier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Suppliers"
                ],
                "summary": "Delete Supplier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Supplier identification",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DeleteSupplierResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "handler.CreateSupplierRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "category_id": {
                    "type": "integer"
                },
                "cnpj": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "service_ids": {
                    "description": "Mudança aqui",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "handler.CreateSupplierResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schemas.SupplierLinkResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.CreateSupplierServiceRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "service_id": {
                    "type": "integer"
                },
                "supplier_id": {
                    "description": "Não defina o ID aqui",
                    "type": "integer"
                }
            }
        },
        "handler.DeleteSupplierResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schemas.SupplierLinkResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.ListSuppliersResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.SupplierLinkResponse"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.ShowSupplierResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schemas.SupplierLinkResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.UpdateSupplierRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "category_id": {
                    "type": "integer"
                },
                "cnpj": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.CreateSupplierServiceRequest"
                    }
                }
            }
        },
        "handler.UpdateSupplierResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schemas.SupplierLinkResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "schemas.ExternalSupplier": {
            "type": "object",
            "properties": {
                "ativo": {
                    "type": "string"
                },
                "bairro": {
                    "type": "string"
                },
                "cep": {
                    "type": "string"
                },
                "cgccfo": {
                    "type": "string"
                },
                "cidade": {
                    "type": "string"
                },
                "codcfo": {
                    "type": "string"
                },
                "codcoligada": {
                    "type": "integer"
                },
                "complemento": {
                    "type": "string"
                },
                "contato": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "nomefantasia": {
                    "type": "string"
                },
                "numero": {
                    "type": "string"
                },
                "rua": {
                    "type": "string"
                },
                "telefone": {
                    "type": "string"
                },
                "tipo": {
                    "type": "string"
                },
                "uf": {
                    "type": "string"
                }
            }
        },
        "schemas.ServiceResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "schemas.SupplierCategory": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "schemas.SupplierLinkResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/schemas.SupplierCategory"
                },
                "category_id": {
                    "type": "integer"
                },
                "cnpj": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "external_supplier": {
                    "$ref": "#/definitions/schemas.ExternalSupplier"
                },
                "id": {
                    "type": "integer"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.ServiceResponse"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}